% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemap_EEZ.R
\name{basemap_EEZ}
\alias{basemap_EEZ}
\title{basmap layer for ggplot with EEZ, Pacific-centered.}
\usage{
basemap_EEZ(
  xlim = c(-30, 330),
  ylim = c(-60, 90),
  south = "up",
  padding = 0.1,
  colour_land = "white",
  colour_ocean = "gray",
  colour_ocean_EEZ = "lightgray",
  colour_border_land = "darkgray",
  colour_border_eez = "#949494",
  colour_lakes = "lightgray"
)
}
\arguments{
\item{xlim}{numerical vector of lenght 2}

\item{ylim}{numerical vector of lenght 2}

\item{south}{description}

\item{colour_land}{defaults to "white",}

\item{colour_ocean}{defaults to "gray",}

\item{colour_ocean_EEZ}{defaults to "lightgray",}

\item{colour_border_land}{defaults to "darkgray",}

\item{colour_border_eez}{defaults to "#949494",}

\item{colour_lakes}{defaults to "lightgray"}
}
\description{
basmap layer for ggplot with EEZ, Pacific-centered.
}
\note{
EEZ data from SOVEREIGN1 layer of Flanders Marine Institute (2019). Maritime Boundaries Geodatabase: Maritime Boundaries and Exclusive Economic Zones (200NM), version 11. Available online at https://www.marineregions.org/. https://doi.org/10.14284/632. File name: World_EEZ_v11_20191118_LR/eez_v11_lowres.shp
}
\author{
Siva Kalyan & Hedvig Skirgård
}
