% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloured_SPLOM.R
\name{coloured_SPLOM}
\alias{coloured_SPLOM}
\title{Creates a colourful SPLOM that is easier to read}
\usage{
coloured_SPLOM(
  df = df,
  pair_colors = "default",
  col_pairs_constraint = "None",
  cols_to_constrain = "None",
  hist_label_size = 3,
  text_cor_size = 5,
  text_strip_size = 12,
  hist_bins = 30,
  cor_value_cut_off = 0.6,
  herringbone = FALSE,
  alpha_point = 0.8
)
}
\arguments{
\item{df}{data-frame with data to be plotted. All columns need to be numerical. All columns will be plotted, so exclude any columns with observation id or other information that is not to be plotted (unless using col_pairs_constraint).}

\item{pair_colors}{character vector. If set to "default", then we use randomcoloR::distinctColorPalette to find a set of distinct colors for the number of plots needed. This argument can also be set to a vector of hex-codes for colors (e.g. c("#E55679", "#5FE3B6", "#D447A0")) and then those specific colors will be used.}

\item{col_pairs_constraint}{characther vector. Default =  "None". Name of column whereby to identify secondary id, see Note for details.}

\item{cols_to_constrain}{characther vector. Default =  "None".Name(s) of column(s) to which the secondary id constraint apply in order to not double-count.}

\item{hist_label_size}{numeric. Default =  3, Font size of the text at the diagonal histograms}

\item{text_cor_size}{numeric. Default = 5. Font size of the text in the upper half reporting the Pearson correlation coefficient.}

\item{text_strip_size}{numeric. Default =  12. Font size of the strips.}

\item{hist_bins}{numeric. Default = 30. Number of bins in the diagonal histograms.}

\item{cor_value_cut_off}{description}

\item{herringbone}{= logical. Default = FALSE. Wether or not to color the SPLOM according to a "herringbone" pattern or unique colors for each pair.}

\item{alpha_point}{numeric. Default = 0.8. Transperency of points in scatterplots.}
}
\description{
Creates a colourful SPLOM that is easier to read
}
\note{
This argument is for when you are comparing variables accross a set of data, but there are two levels of granularity to defining the observation. For example, in the multilingual corpora collection Universal Dependencies (UD) there are datasets that can be tied to languages (many-to-one mapping) and in Grambank & Google's Linguameta there are only languages. If we want to compare information from UD, Grambank and Linguameta, we don't want to double-count the same languages when comparing Grambank and Linguameta. Therefore, in such an instance, the user should specify a secondary ID variable (e.g. "glottocode" in col_pairs_constraint) and the names of the columns (variable) to which this applies (e.g. "Grambank Fusion" and "Google Pop") in cols_to_constrain.
}
