% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit_packages.R
\name{credit_packages}
\alias{credit_packages}
\title{Lists used packages in a set of R-script and generates citation keys, bibtex-file, tables etc.}
\usage{
credit_packages(
  fns = NULL,
  extra_pkgs = NULL,
  output_dir = NULL,
  print_bibTeX = TRUE,
  print_tex_citation_string = TRUE,
  compare_loaded_with_used = TRUE,
  report_most_used_pkgs = TRUE,
  print_LaTeX_table = TRUE,
  print_tsv = TRUE,
  report_script_with_most_funs = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{fns}{character vector of file-names to R-scripts to read in.}

\item{extra_pkgs}{character vector. Names of R-packages, either instead of fns or in addition to it. Sometimes some packages are missed when searching through the scripts, if you add their names here they'll be included in the output despite this.}

\item{output_dir}{character vector. Name of directory to print bibTeX file and citation-keys. Necessary if print_bibTeX == TRUE and/or print_tex_citation_string == TRUE.}

\item{print_bibTeX}{logical. If TRUE, a bibTeX file is written with entries for the packages found to be used. File will be written to output_dir as "used_pkgs.bib"}

\item{print_tex_citation_string}{logical. If TRUE, a txt-file will be written with a string of all the TeX citation keys in a string. This txt-file can be included in a TeX document with "\\input{}". It will be in output_dir with the name "citation_keys.txt".}

\item{compare_loaded_with_used}{logical. If TRUE, the set of packages that the scripts used are compared to those loaded currently in the environment and reported in terminal.}

\item{report_most_used_pkgs}{logical. If TRUE, the function will report which are the 5 packages you use the most functions from to the terminal.}

\item{print_LaTeX_table}{logical. If TRUE, a LaTeX table will be rendered with each pagkage as a row and a column for version loaded.}

\item{print_tsv}{logical. If TRUE, a table of packages will be printed to a tsv-file in the output_dir, with a column for version loaded.}

\item{report_script_with_most_funs}{logical. If TRUE, the function will report which script has the most function calls to the terminal.}

\item{verbose}{logical. If TRUE, the function will be more talkative.}
}
\value{
Data-frame of all used functions. Depending on the arguments, the function also returns output to the terminal and/or files written to the output directory.
}
\description{
Lists used packages in a set of R-script and generates citation keys, bibtex-file, tables etc.
}
\note{
In cases where it is not clear which specific package a function is from several packages are returned for that function. This can for example be the case with filter(). This function uses an adapted version of 'list.functions.in.file' from the package NC.misc. That function was written by Nicholas Cooper 'njcooper (at) gmx.co.uk'. The adjustment consist of removing unique() such that all instances are reported.
}
\author{
Hedvig Skirgård
}
