% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_as_tip.R
\name{.fun_Descendants_node_labels}
\alias{.fun_Descendants_node_labels}
\title{Prunes a tree down to just the tips that are specified. If the specified labels belong to node-labels, all tips below the node are pruned away until the node becomes a tip.}
\usage{
.fun_Descendants_node_labels(tree, node, type = "tips")
}
\arguments{
\item{tree}{phylo-class object}

\item{tips_and_nodes_to_keep}{character vector of tips and/or node labels that are to be kept}
}
\description{
Prunes a tree down to just the tips that are specified. If the specified labels belong to node-labels, all tips below the node are pruned away until the node becomes a tip.
}
\note{
These function was written to solve a specific problem. The functions are written perhaps a bit unusually, with tidyverse pipes and data.frames where more "elegant" base-solutions are available. This function was written for specific use cases I have encountered. If you find them useful and would like them to be improved in terms of functionality, elegance and/or speed let me know. For another take on this problem, see Erich Round's solution here: https://github.com/erichround/glottoTrees/blob/99338ff3e62fa3f282eb19ed55a60c564fd843bc/R/topology.R#L326
}
\author{
Hedvig Skirgård
}
