% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_lines.R
\name{fetch_lines}
\alias{fetch_lines}
\title{Fetch text content from internet or other location and write to new location.}
\usage{
fetch_lines(
  url = NULL,
  out_dir = NULL,
  add_date_fetched = TRUE,
  overwrite = FALSE,
  verbose = FALSE,
  comment_char = "#"
)
}
\arguments{
\item{url}{character vector. File-path or url pointing to a text file.}

\item{out_dir}{character vector. File-path of folder where the file is to be copied to}

\item{overwrite}{logical. If TRUE, the function will overwrite the file if it already exists.}

\item{verbose}{logical If TRUE, function will message.}

\item{comment_char}{character vector. Specifies character used for comments in code. Defaults to '#'.}

\item{add_date_fetch}{logical. If TRUE, the file will have an added line to the top that indicates when it was fetched and from where}
}
\description{
Fetch text content from internet or other location and write to new location.
}
\examples{
# fetch_lines(url = "https://github.com/HedvigS/R_grambank_cookbook/raw/main/functions/make_theo_scores.R", out_dir = ".")

}
\author{
Hedvig Skirgård
}
