% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zenodo_dir.R
\name{get_zenodo_dir}
\alias{get_zenodo_dir}
\title{Downloads content from Zenodo based on a download-URL, places it in a given spot on the local machine. Back-up function for rcldf::cldf. Can be used on any Zenodo-record, not just CLDF.}
\usage{
get_zenodo_dir(url, exdir, drop_dir_level = TRUE)
}
\arguments{
\item{url}{character vector. URL string, e..g "https://zenodo.org/record/7740822/files/grambank/grambank-analysed-v1.0.zip"}

\item{exdir}{character vector. name of folder where the files are to be written to}

\item{drop_dir_level}{logical. If FALSE, a dir will exist inside the exdir that bears the name from the zip-file from Zenodo. If TRUE, this dir-evel will be removed, the contents of the zipped-file will be placed directly in exdir.}
}
\description{
Downloads content from Zenodo based on a download-URL, places it in a given spot on the local machine. Back-up function for rcldf::cldf. Can be used on any Zenodo-record, not just CLDF.
}
\author{
Hedvig Skirgård
}
