% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLarbor_make_CLarbor_df.R
\name{make_CLarbor_df}
\alias{make_CLarbor_df}
\title{Convert a minimum spanning arborescence into a tidy data frame}
\usage{
make_CLarbor_df(weighted_CLarbor)
}
\arguments{
\item{weighted_CLarbor}{An \code{igraph} weighted tree output by \code{\link[=chuliu]{chuliu()}}.}
}
\value{
A data frame with one row per edge, and with the following columns:
\itemize{
\item \strong{Source feature}: The label of the source node of the edge.
\item \strong{Target feature}: The label of the target node of the edge.
\item \strong{Weight}: The weight of the edge.
\item \strong{Dependency}: 1 minus the weight of the edge, corresponding to dependency
strength in the context of this package.
}
}
\description{
Takes a Chu-Liu/Edmonds arborescence output by \code{\link[=chuliu]{chuliu()}} and converts it
into a data frame with one row per edge.
}
