% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.d_wrapper.R
\name{phylo.d_wrapper}
\alias{phylo.d_wrapper}
\title{Wrapped function to caper::phylo.d which performs some important sanity checks.}
\usage{
phylo.d_wrapper(data, var.name, ...)
}
\arguments{
\item{data}{A list of the class "comparative.data". Created by \code{caper::comparative.data}.}

\item{var.name}{Character vector. The name of the variable in \code{data} holding the binary variable of interest. This will be passed to \code{phylo.d} and passed (as a symbol) to the argument \code{binvar}.}

\item{...}{Additional parameters passed to \code{phylo.d}.}
}
\value{
If no error occurs, the function returns an object of class 'phylo.d' just as \code{caper::phylo.d}. See \code{?caper::phylo.d} for more details. If the wrapper-function finds problem, it will trigger error messages.
}
\description{
Wrapped function to caper::phylo.d which performs some important sanity checks.
}
\author{
wrapper: Hedvig Skirgård and Siva Kalyan; original caper::phylo.d-function: Susanne Fritz and David Orme.
}
