% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SH.misc-package.R
\docType{package}
\name{SH.misc-package}
\alias{SH.misc}
\alias{SH.misc-package}
\title{SH.misc: Assorted Functions for Data Science in Linguistics}
\description{
Functions for manipulating tabular data, inferring typological feature dependencies, and creating Pacific-centred world maps with Exclusive Economic Zones (EEZs).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/HedvigS/SH.misc}
  \item Report bugs at \url{https://github.com/HedvigS/SH.misc/issues}
}

}
\author{
\strong{Maintainer}: Hedvig Skirgård \email{hedvig.public@gmail.com} (\href{https://orcid.org/0000-0002-7748-2381}{ORCID})

Authors:
\itemize{
  \item Siva sivak.au@gmail.com Kalyan (\href{https://orcid.org/0000-0002-6302-6641}{ORCID})
}

}
\keyword{internal}
