% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_delim.R
\name{stack_delim}
\alias{stack_delim}
\title{Stacks content of multiple tsv-files on-top of each other.}
\usage{
stack_delim(fns = fns, verbose = TRUE, delim = "\\t")
}
\arguments{
\item{fns}{character vector of file names.}

\item{verbose}{logical. If TRUE, function will report on which file it is at etc.}

\item{delim}{character vector for delim, defaults to "\\t".}
}
\value{
data-frame with content of all tsvs stacked and a column with filename.
}
\description{
Stacks content of multiple tsv-files on-top of each other.
}
\note{
All content will be turned into character to facilitate joining and then the col-types will be converted back to appropriate class for the whole data-frame.
}
\author{
Hedvig Skirgård
}
