% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_tsvs.R
\name{stack_tsvs}
\alias{stack_tsvs}
\title{Stacks content of multiple tsv-files on-top of each other.}
\usage{
stack_tsvs(fns = fns, verbose = TRUE)
}
\arguments{
\item{fns}{character vector of file names.}

\item{verbose}{logical. If TRUE, function will report on which file it is at etc.}
}
\value{
data-frame with content of all tsvs stacked and a column with filename.
}
\description{
Stacks content of multiple tsv-files on-top of each other.
}
\note{
All content will be turned into character to facilitate joining and then the col-types will be converted back to appropriate class for the whole data-frame.
}
\author{
Hedvig Skirgård.
}
