% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/CLarbor_CLarbor_from_dep_matrix_check_all_roots.R
\name{CLarbor_from_dep_matrix_check_all_roots}
\alias{CLarbor_from_dep_matrix_check_all_roots}
\title{Applies the function CLarbor::from_dep_matrix_known_root to each node in the graph, then evaluates which one was indeed the best root.}
\usage{
CLarbor_from_dep_matrix_check_all_roots(dep_matrix = NULL, return.all = FALSE)
}
\arguments{
\item{dep_matrix}{the output of the function dependency_matrix_cond_MI()}

\item{return.all}{logical. If TRUE, all possible trees/forests are returend. If FALSE, only the best is returned. Defaults to FALSE.}
}
\description{
Applies the function CLarbor::from_dep_matrix_known_root to each node in the graph, then evaluates which one was indeed the best root.
}
\note{
We would like to thank Harald Hammarström for his invaluable guidance and support in the creation of these functions. See Hammarström, H., & O’Connor, L. (2013). Dependency-sensitive typological distance. In Approaches to measuring linguistic differences. De Gruyter.
}
\author{
Siva Kalyan and Hedvig Skirgård
}
