% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLarbor_CLarbor_from_dep_matrix_known_root.R
\name{CLarbor_from_dep_matrix_known_root}
\alias{CLarbor_from_dep_matrix_known_root}
\title{Prunes the dependency graph according to the Chu-Liu/Edmonds algorithm given a particular node as the root.}
\usage{
CLarbor_from_dep_matrix_known_root(dep_matrix = NULL, root)
}
\arguments{
\item{dep_matrix}{the output of the function dependency_matrix_cond_MI()}

\item{roor}{.}
}
\description{
Prunes the dependency graph according to the Chu-Liu/Edmonds algorithm given a particular node as the root.
}
\note{
We would like to thank Harald Hammarström for his invaluable guidance and support in the creation of these functions. See Hammarström, H., & O’Connor, L. (2013). Dependency-sensitive typological distance. In Approaches to measuring linguistic differences. De Gruyter.
}
\author{
Siva Kalyan and Hedvig Skirgård
}
